Refactor logging to use getLogger() instead of qDebug() and qWarning()
